requirements.txt
aiohttp==3.5.1
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
PyYAML==3.13
telepot==12.7
urllib3==1.24.1
yarl==1.3.0
